a step in the right direction
I found out that the culprit for moveMapPrevious not working was that I was fetching the same coordinates I was on. Fixed this dumb bug, I need a way to let moveMapPrevious know what coordinates moveMapNext is using so that it can substract or do something with the list to iterate it backwards.