bug and code clarity fixes
Correct logic for all the different components based on the current state of the program.

Added a componentsSetEnabled to unload some code to a small method for better clarity

moved the delete event button for redeability.

now deleteButton is also triggered by pressing the ESC key on the user's keyboard